changed form long long to ints to store values

+++ include/compact_trie.hpp
@@ -88,8 +88,8 @@ class CompactTrie: public TrieInterface{
             Recives index in bit vector
             Returns index of next 0
         */
-        uint64_t succ0(uint64_t it){
-            uint64_t cant_0 = b_rank0(it);
+        uint32_t succ0(uint32_t it){
+            uint32_t cant_0 = b_rank0(it);
             return b_sel0(cant_0 + 1);
         }
         
@@ -97,8 +97,8 @@ class CompactTrie: public TrieInterface{
             Recives index in bit vector
             Returns index of previous 0
         */
-        uint64_t prev0(uint64_t it){
-            uint64_t cant_0 = b_rank0(it);
+        uint32_t prev0(uint32_t it){
+            uint32_t cant_0 = b_rank0(it);
             return b_sel0(cant_0);
         }
 
@@ -106,7 +106,7 @@ class CompactTrie: public TrieInterface{
             Recives index of current node and the child that is required
             Returns index of the nth child of current node
         */
-        uint64_t child(uint64_t it, uint64_t n){
+        uint32_t child(uint32_t it, uint32_t n){
             return b_sel0(b_rank1(it+n)) + 1;
         }
 
@@ -114,7 +114,7 @@ class CompactTrie: public TrieInterface{
             Recives index of node whos children we want to count
             Returns how many children said node has
         */
-        uint64_t childrenCount(uint64_t it){
+        uint32_t childrenCount(uint32_t it){
             return succ0(it) - it;
         }
 
@@ -122,7 +122,7 @@ class CompactTrie: public TrieInterface{
             Recives node index
             Returns index of position in parent
         */
-        uint64_t getPosInParent(uint64_t it){
+        uint32_t getPosInParent(uint32_t it){
             return b_sel1(b_rank0(it));
         }
 
@@ -130,8 +130,8 @@ class CompactTrie: public TrieInterface{
             Recives index of node
             Return which child of its parent it is
         */
-        uint64_t childRank(uint64_t it){
-            uint64_t pos = getPosInParent(it);
+        uint32_t childRank(uint32_t it){
+            uint32_t pos = getPosInParent(it);
             return pos - prev0(pos);
         }
 
@@ -139,22 +139,22 @@ class CompactTrie: public TrieInterface{
             Recives index of node
             Returns index of parent node
         */  
-        uint64_t parent(uint64_t it){
-            uint64_t pos = getPosInParent(it);
+        uint32_t parent(uint32_t it){
+            uint32_t pos = getPosInParent(it);
             return prev0(pos) + 1;
         }
 
         /*
             Returns key that corresponds to given node(it)
         */
-        uint64_t key_at(uint64_t it){
+        uint32_t key_at(uint32_t it){
             return wt[b_rank0(it)-2];
         }
 
         /*
             Returns i-th element of the original sequence s 
         */
-        uint64_t get_wt_at(uint64_t i){
+        uint32_t get_wt_at(uint32_t i){
             return wt[i];
         }
 
@@ -180,7 +180,7 @@ class CompactTrie: public TrieInterface{
             initializeSupport();
         }
         
-        uint64_t getMaxSequence(){
+        uint32_t getMaxSequence(){
             return 0;
         }
 };
