+++ include/iterator.hpp
@@ -15,8 +15,8 @@ class Iterator{
         // Iterator(CompactTrie* ct){};
         virtual ~Iterator(){};
         virtual bool atEnd() = 0;
-        virtual uint64_t key() = 0;
-        virtual void seek(uint64_t seekKey) = 0;
+        virtual uint32_t key() = 0;
+        virtual void seek(uint32_t seekKey) = 0;
         virtual void next() = 0;
         virtual void open() = 0;
         virtual void up() = 0;
@@ -30,7 +30,7 @@ class Iterator{
         //Temporal
         virtual void getIteratorPos() = 0;
         virtual TrieInterface* getCompactTrie() = 0;
-        virtual uint64_t getTuple() = 0;
+        virtual uint32_t getTuple() = 0;
 };
 
 #endif
